Foldeed CubeMeshModel.h/.cpp into MeshModel.h/.cpp
