+++ CG_skel_w_MFC/MeshModel.h
@@ -8,8 +8,11 @@
 #include "Renderer.h"
 #include "Geometry.h"
 
-using namespace std;
+#define CUBE_DEFAULT_SIZE 20
+#define CUBE_MESH_DEFAULT_MIN (vec3(0, 0, 0))
+#define CUBE_MESH_DEFAULT_MAX (vec3(CUBE_DEFAULT_SIZE, CUBE_DEFAULT_SIZE, CUBE_DEFAULT_SIZE))
 
+using namespace std;
 
 struct FaceIdcs
 {
@@ -67,4 +70,23 @@ public:
 	void translate(vec3 translation);
 	void rotate(vec3 axis, float angle);
 	void scale(vec3 scale);
+	// MeshModel options
+	void drawBoundingBox(Renderer& renderer);
+	void drawVertexNormals(Renderer& renderer);
+	void drawFaceNormals(Renderer& renderer);
+};
+
+class CubeMeshModel : public MeshModel
+{
+public:
+	CubeMeshModel() noexcept;
+};
+
+class BoxMeshModel : public MeshModel
+{
+public:
+	BoxMeshModel();
+	BoxMeshModel(vec3 min, vec3 max);
+	static BoxMeshModel makeBoundingBox(float red, float green, float blue,
+		vec3 min, vec3 max);
 };
